+++ CLF_Queue/CLF_Queue/TC_LFQueue64.h
@@ -184,7 +184,7 @@ namespace procademy
 	{
 		DWORD i;
 		Node* pHead = GetNodeAddress(mHead)->next;
-		int curSize = mSize;
+		DWORD curSize = (DWORD)mSize;
 		size = size > curSize ? curSize : size;
 
 		for (i = 0; i < size; ++i)
